versioned API link improvements (#26)
Add API version to all group and operation anchor links, so that when you click to expand a group or operation, the browser URL will always show the version-specific link, ready for copying or sending to others.  Unversioned links are still supported, but are not generated anymore.
Selecting an api version via the dropdown list will now update the url hash, and will attempt to preserve the currently selected group and/or endpoint across the version switch.